+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `--recreate` flag for `git clone` command that recreates previously cloned repository
+
 ## [0.3.0] - 2022-07-28
 
 ### Added